additional enhancements of feedback and colorloop feature
1.2 streamlines some of the methods on the # side by pulling the entire
JArray in from room/group actions, I forsee additional work in this area
to deploy an event broker solution to update all the affected room
modules and light modules.

1.2 also adds the color loop flag in SIMPL, so if this is latched on a
room/group on cmd will cycle through the various hue's until the group
is turned off. I suspect it would be better to simply on rising edge
trigger the group/room on with the colorloop effect and on the trailing
edge turn the load off or remember the last on/off state when the loop
was triggered to go back to that original hue. perhaps the next rev.

The Light bulb individual modules need to be reviewed and updated, not
sure how useful they are as I typically control groups not individual
loads. the problem is of course how do you define each bulb on the SIMPL
side, currently it's the UID, open to suggestions.