get bridge info enhancements
re-worked the bulbs/groups/scenes routines to utilize a better event
based communication system. This change seemed necessary after I
discovered my house network was a little sluggish with Sonos UDP traffic
and the JSON responses weren't being processed properly because of the
delay.  The Processor module was updated and the CLZ has been refreshed.
One Issue I am aware of and will implement a fix soon is when a Hue room
has been defined with no loads, this will cause an exception and bomb
out the module.